debug: changed CastAdapter.kt, fragment_movie.xml * CastAdapter.kt changed from RecyclerView.Adapter to ListAdapter * fragment_movie.xml changed from ConstraintLayout to FrameLayout * item_cast.xml changed from ConstraintLayout to FrameLayout
