+++ .idea/misc.xml
@@ -1,5 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="DesignSurface">
+    <option name="filePathToZoomLevelMap">
+      <map>
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/activity_main.xml" value="0.264" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/fragment_movie.xml" value="0.1" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/item_cast.xml" value="0.536" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/item_header.xml" value="0.25416666666666665" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/item_placeholder_recycler.xml" value="0.25416666666666665" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/layout/merge_fragment_movie.xml" value="0.25416666666666665" />
+        <entry key="..\:/Users/Flyin/AndroidStudioProjects/MoviesFinder/app/src/main/res/menu/bottom_navigation_bar.xml" value="0.26510416666666664" />
+      </map>
+    </option>
+  </component>
   <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
