+++ app/src/main/java/com/lefarmico/moviesfinder/adapter/CastAdapter.kt
@@ -1,19 +1,21 @@
-package com.lefarmico.moviesfinder.adapters
+package com.lefarmico.moviesfinder.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.lefarmico.moviesfinder.R
 import com.lefarmico.moviesfinder.data.appEntity.Cast
 import com.lefarmico.moviesfinder.databinding.ItemCastBinding
+import com.lefarmico.moviesfinder.extension.loadWithThemeParams
 import com.lefarmico.moviesfinder.private.ApiConstants
 import com.squareup.picasso.Picasso
 
-class CastAdapter : RecyclerView.Adapter<CastAdapter.ViewHolder>() {
-
-    private lateinit var castList: List<Cast>
+class CastAdapter : ListAdapter<Cast, CastAdapter.ViewHolder>(
+    CastDuffUtilCallback()
+) {
 
     class ViewHolder(itemCastBinding: ItemCastBinding) : RecyclerView.ViewHolder(itemCastBinding.root) {
         private val personName: TextView = itemCastBinding.personName
@@ -26,13 +28,21 @@ class CastAdapter : RecyclerView.Adapter<CastAdapter.ViewHolder>() {
 
             Picasso
                 .get()
-                .load(ApiConstants.IMAGES_URL + "w342" + cast.profilePath)
-                .fit()
-                .error(R.drawable.ic_launcher_foreground)
-                .placeholder(R.drawable.ic_launcher_foreground)
+                .loadWithThemeParams(ApiConstants.IMAGES_URL + "w342" + cast.profilePath)
                 .into(poster)
         }
     }
+
+    private class CastDuffUtilCallback : DiffUtil.ItemCallback<Cast>() {
+        override fun areItemsTheSame(oldItem: Cast, newItem: Cast): Boolean {
+            return oldItem.personId == newItem.personId
+        }
+
+        override fun areContentsTheSame(oldItem: Cast, newItem: Cast): Boolean {
+            return oldItem == newItem
+        }
+    }
+
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
         ViewHolder(
             ItemCastBinding.inflate(
@@ -41,13 +51,7 @@ class CastAdapter : RecyclerView.Adapter<CastAdapter.ViewHolder>() {
         )
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(castList[position])
-    }
-
-    override fun getItemCount(): Int = castList.size
-
-    fun setItems(list: List<Cast>) {
-        castList = list
-        notifyDataSetChanged()
+        val castItem = getItem(position)
+        holder.bind(castItem)
     }
 }
