+++ app/src/main/java/com/lefarmico/moviesfinder/adapter/ItemAdapter.kt
@@ -1,4 +1,5 @@
-package com.lefarmico.moviesfinder.adapters
+
+package com.lefarmico.moviesfinder.adapter
 
 import android.util.Log
 import android.view.LayoutInflater
@@ -7,30 +8,23 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.lefarmico.moviesfinder.R
-import com.lefarmico.moviesfinder.customViews.RatingView
 import com.lefarmico.moviesfinder.data.appEntity.ItemHeader
-import com.lefarmico.moviesfinder.databinding.ItemWatchListRecyclerBinding
+import com.lefarmico.moviesfinder.databinding.ItemBinding
 import com.lefarmico.moviesfinder.private.ApiConstants
-import com.lefarmico.moviesfinder.utils.RecyclerViewAdapterWithListener
 import com.squareup.picasso.Picasso
 
-class WatchListRecyclerViewAdapter : RecyclerViewAdapterWithListener<ItemHeader, WatchListRecyclerViewAdapter.ViewHolder>() {
+class ItemAdapter(
+    private val listener: (ItemHeader) -> Unit
+) : RecyclerView.Adapter<ItemAdapter.ViewHolder>() {
 
-    override var items = mutableListOf<ItemHeader>()
+    private var items = mutableListOf<ItemHeader>()
 
-    override var onClickEvent: OnClickEvent<ItemHeader>? = null
-
-    class ViewHolder(itemBinding: ItemWatchListRecyclerBinding) :
-        RecyclerView.ViewHolder(itemBinding.root) {
+    class ViewHolder(itemBinding: ItemBinding) : RecyclerView.ViewHolder(itemBinding.root) {
         private val poster: ImageView = itemBinding.poster
-        val title: TextView = itemBinding.title
-        val description: TextView = itemBinding.description
-        val rating: RatingView = itemBinding.movieRate
+        val title: TextView = itemBinding.movieTitle
 
         fun bind(itemHeader: ItemHeader) {
             title.text = itemHeader.title
-            description.text = itemHeader.description
-            rating.setRatingValue(itemHeader.rating)
             Picasso
                 .get()
                 .load(ApiConstants.IMAGES_URL + "w342" + itemHeader.posterPath)
@@ -40,24 +34,27 @@ class WatchListRecyclerViewAdapter : RecyclerViewAdapterWithListener<ItemHeader,
                 .into(poster)
         }
     }
-    override fun onCreateViewHolderWithListener(
-        parent: ViewGroup,
-        viewType: Int
-    ): ViewHolder {
-        val inlfater = LayoutInflater.from(parent.context)
-        val binding = ItemWatchListRecyclerBinding.inflate(inlfater, parent, false)
-        return ViewHolder(binding)
-    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
+        ViewHolder(
+            ItemBinding.inflate(
+                LayoutInflater.from(parent.context), parent, false
+            )
+        )
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         Log.d("TAG", "bind, position = $position")
         holder.bind(items[position])
+        holder.itemView.setOnClickListener {
+            listener(items[position])
+        }
     }
 
     override fun getItemCount(): Int = items.size
 
-    fun setItemHeaders(itemsList: List<ItemHeader>) {
+    fun setItems(itemsList: List<ItemHeader>) {
         items.addAll(itemsList)
+        // TODO Change it
         notifyDataSetChanged()
     }
 
@@ -65,6 +62,7 @@ class WatchListRecyclerViewAdapter : RecyclerViewAdapterWithListener<ItemHeader,
         val updateList = mutableListOf<ItemHeader>()
         updateList.addAll(itemsList)
         items = updateList
+        // TODO Change it
         notifyDataSetChanged()
     }
