+++ app/src/main/java/com/lefarmico/moviesfinder/adapter/SpinnerProviderAdapter.kt
@@ -1,4 +1,4 @@
-package com.lefarmico.moviesfinder.adapters
+package com.lefarmico.moviesfinder.adapter
 
 import android.content.Context
 import android.view.KeyEvent
