+++ app/src/main/java/com/lefarmico/moviesfinder/adapter/WatchListRecyclerViewAdapter.kt
@@ -1,5 +1,4 @@
-
-package com.lefarmico.moviesfinder.adapters
+package com.lefarmico.moviesfinder.adapter
 
 import android.util.Log
 import android.view.LayoutInflater
@@ -8,23 +7,30 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.lefarmico.moviesfinder.R
+import com.lefarmico.moviesfinder.customViews.RatingView
 import com.lefarmico.moviesfinder.data.appEntity.ItemHeader
-import com.lefarmico.moviesfinder.databinding.ItemBinding
+import com.lefarmico.moviesfinder.databinding.ItemWatchListRecyclerBinding
 import com.lefarmico.moviesfinder.private.ApiConstants
+import com.lefarmico.moviesfinder.utils.RecyclerViewAdapterWithListener
 import com.squareup.picasso.Picasso
 
-class ItemAdapter(
-    private val listener: (ItemHeader) -> Unit
-) : RecyclerView.Adapter<ItemAdapter.ViewHolder>() {
+class WatchListRecyclerViewAdapter : RecyclerViewAdapterWithListener<ItemHeader, WatchListRecyclerViewAdapter.ViewHolder>() {
 
-    private var items = mutableListOf<ItemHeader>()
+    override var items = mutableListOf<ItemHeader>()
 
-    class ViewHolder(itemBinding: ItemBinding) : RecyclerView.ViewHolder(itemBinding.root) {
+    override var onClickEvent: OnClickEvent<ItemHeader>? = null
+
+    class ViewHolder(itemBinding: ItemWatchListRecyclerBinding) :
+        RecyclerView.ViewHolder(itemBinding.root) {
         private val poster: ImageView = itemBinding.poster
-        val title: TextView = itemBinding.movieTitle
+        val title: TextView = itemBinding.title
+        val description: TextView = itemBinding.description
+        val rating: RatingView = itemBinding.movieRate
 
         fun bind(itemHeader: ItemHeader) {
             title.text = itemHeader.title
+            description.text = itemHeader.description
+            rating.setRatingValue(itemHeader.rating)
             Picasso
                 .get()
                 .load(ApiConstants.IMAGES_URL + "w342" + itemHeader.posterPath)
@@ -34,26 +40,25 @@ class ItemAdapter(
                 .into(poster)
         }
     }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder =
-        ViewHolder(
-            ItemBinding.inflate(
-                LayoutInflater.from(parent.context), parent, false
-            )
-        )
+    override fun onCreateViewHolderWithListener(
+        parent: ViewGroup,
+        viewType: Int
+    ): ViewHolder {
+        val inlfater = LayoutInflater.from(parent.context)
+        val binding = ItemWatchListRecyclerBinding.inflate(inlfater, parent, false)
+        return ViewHolder(binding)
+    }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         Log.d("TAG", "bind, position = $position")
         holder.bind(items[position])
-        holder.itemView.setOnClickListener {
-            listener(items[position])
-        }
     }
 
     override fun getItemCount(): Int = items.size
 
-    fun setItems(itemsList: List<ItemHeader>) {
+    fun setItemHeaders(itemsList: List<ItemHeader>) {
         items.addAll(itemsList)
+        // TODO Change it
         notifyDataSetChanged()
     }
 
@@ -61,6 +66,7 @@ class ItemAdapter(
         val updateList = mutableListOf<ItemHeader>()
         updateList.addAll(itemsList)
         items = updateList
+        // TODO Change it
         notifyDataSetChanged()
     }
