+++ app/src/main/java/com/lefarmico/moviesfinder/extension/PicassoExtension.kt
@@ -1,0 +1,13 @@
+package com.lefarmico.moviesfinder.extension
+
+import com.lefarmico.moviesfinder.R
+import com.squareup.picasso.Picasso
+import com.squareup.picasso.RequestCreator
+
+fun Picasso.loadWithThemeParams(path: String?): RequestCreator {
+    return this
+        .load(path)
+        .fit()
+        .error(R.drawable.ic_launcher_foreground)
+        .placeholder(R.drawable.ic_launcher_foreground)
+}
