+++ app/src/main/java/com/lefarmico/moviesfinder/viewModels/ProfileFragmentViewModel.kt
@@ -3,7 +3,7 @@ package com.lefarmico.moviesfinder.viewModels
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.lefarmico.moviesfinder.App
-import com.lefarmico.moviesfinder.adapters.WatchListRecyclerViewAdapter
+import com.lefarmico.moviesfinder.adapter.WatchListRecyclerViewAdapter
 import com.lefarmico.moviesfinder.data.Interactor
 import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.schedulers.Schedulers
