+++ app/src/main/res/drawable/bottom_navigation_selector.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
     <item android:color="@color/active_color" android:state_checked="true"/>
-    <item android:color="@color/main_contrast_color" android:state_checked="false"/>
+    <item android:color="@color/on_background_color" android:state_checked="false"/>
 </selector>
