+++ app/src/main/res/drawable/genre_item_shape.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <stroke
-        android:color="@color/main_contrast_color"
+        android:color="@color/on_background_color"
         android:width="1dp"/>
     <corners android:radius="3dp"/>
 </shape>
