+++ app/src/main/res/layout/bottom_sheet_movie_details.xml
@@ -109,13 +109,13 @@
             android:textSize="12sp"
             style="@style/TextStyle"
             android:text="@string/genres"
-            android:textColor="@color/main_contrast_color"
+            android:textColor="@color/on_background_color"
             android:orientation="horizontal"/>
         <TextView
             android:id="@+id/length"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textColor="@color/main_contrast_color"
+            android:textColor="@color/on_background_color"
             android:textSize="12sp"
             style="@style/TextStyle"
             android:text="@string/duration"/>
@@ -123,7 +123,7 @@
             android:id="@+id/release_date"
             android:textSize="12sp"
             style="@style/TextStyle"
-            android:textColor="@color/main_contrast_color"
+            android:textColor="@color/on_background_color"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/release_date"/>
@@ -172,7 +172,7 @@
                     app:layout_constraintStart_toStartOf="@id/movie_rate"
                     app:layout_constraintEnd_toEndOf="@id/movie_rate"
                     app:layout_constraintBottom_toBottomOf="parent"
-                    android:textColor="@color/main_contrast_color"
+                    android:textColor="@color/on_background_color"
                     android:text="@string/users_rate"/>
 
                 <View
@@ -203,7 +203,7 @@
                     app:layout_constraintStart_toStartOf="@id/watched_list_count"
                     app:layout_constraintEnd_toEndOf="@id/watched_list_count"
                     app:layout_constraintBottom_toBottomOf="parent"
-                    android:textColor="@color/main_contrast_color"
+                    android:textColor="@color/on_background_color"
                     android:text="@string/my_rate"/>
 
                 <View
@@ -236,7 +236,7 @@
                     app:layout_constraintStart_toStartOf="@id/favorite_toggle_button"
                     app:layout_constraintEnd_toEndOf="@id/favorite_toggle_button"
                     app:layout_constraintBottom_toBottomOf="parent"
-                    android:textColor="@color/main_contrast_color"
+                    android:textColor="@color/on_background_color"
                     android:text="@string/watch_list"/>
             </androidx.constraintlayout.widget.ConstraintLayout>
