+++ app/src/main/res/layout/item_provider_spinner_default.xml
@@ -19,7 +19,7 @@
         android:layout_height="24dp"
         android:src="@drawable/ic_baseline_keyboard_arrow_down_24"
         android:layout_marginEnd="24dp"
-        app:tint="@color/main_contrast_color"
+        app:tint="@color/on_background_color"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
