+++ app/src/main/res/layout/merge_fragment_movie.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
     <androidx.recyclerview.widget.RecyclerView
@@ -11,7 +10,7 @@
         android:layout_height="wrap_content"
         android:visibility="visible"
         tools:ignore="RtlSymmetry"
-        app:layout_constraintTop_toTopOf="parent"/>
+        />
 
     <ProgressBar
         android:id="@+id/progress_bar"
@@ -19,7 +18,6 @@
         android:layout_height="wrap_content"
         android:visibility="visible"
         tools:ignore="RtlSymmetry"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"/>
+        android:layout_gravity="center"/>
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+</FrameLayout>
