+++ app/src/main/res/values-night/colors.xml
@@ -9,7 +9,7 @@
     <color name="white">#FFFFFF</color>
 
     <!-- push elements colors -->
-    <color name="active_color">#FFD600</color>
+    <color name="active_color">#FFCA28</color>
     <color name="passive_color">#F0F0F3</color>
 
     <!-- menu colors -->
@@ -23,7 +23,7 @@
     <color name="main_color">#171722</color>
     <color name="main_color_dark">#0D0D13</color>
 
-    <color name="main_contrast_color">#F1F0F5</color>
+    <color name="on_background_color">#F1F0F5</color>
 
     <color name="bottom_sheet_button_color">#222232</color>
 </resources>
