+++ app/src/main/res/values-night/themes.xml
@@ -1,5 +1,5 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <style name="DarkYellowTheme" parent="Theme.AppCompat.NoActionBar">
+<resources>
+    <style name="DarkYellowTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">@color/main_color</item>
         <item name="colorPrimaryDark">@color/main_color_dark</item>
@@ -11,7 +11,7 @@
         <item name="colorAccent">@color/active_color</item>
 
         <!-- Status bar color. -->
-        <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimary</item>
+        <item name="android:statusBarColor" >?attr/colorPrimary</item>
         <item name="android:colorBackground">@color/menu_background</item>
         <!-- Customize your theme here. -->
