+++ app/src/main/res/values/styles.xml
@@ -7,7 +7,7 @@
 
     <style name="ToolBarStyle" parent="DarkYellowTheme">
         <item name="android:background">@color/main_color</item>
-        <item name="titleTextColor">@color/main_contrast_color</item>
+        <item name="titleTextColor">@color/on_background_color</item>
     </style>
 
     <style name="BottomNavigatorStyle">
@@ -21,6 +21,22 @@
         <item name="android:textSize">14sp</item>
         <item name="android:typeface">sans</item>
         <item name="android:fontFamily">@font/roboto_regular</item>
+    </style>
+
+    <style name="TextStyle.Body">
+        <item name="android:textStyle">normal</item>
+        <item name="android:textSize">12sp</item>
+        <item name="android:typeface">sans</item>
+        <item name="android:fontFamily">@font/roboto_regular</item>
+    </style>
+
+    <style name="TextStyle.Subtitle">
+        <item name="android:textSize">12sp</item>
+    </style>
+
+    <style name="TextStyle.Subtitle.Contrast">
+        <item name="android:textSize">12sp</item>
+        <item name="android:textColor">@color/white</item>
     </style>
 
     <style name="TextStyle.Warning">
